Remove pytest as an install dependency
