+++ setup.cfg
@@ -21,7 +21,7 @@ select=F, E101, E111, E112, E113, E401, E402, E711, E722 --max-line-length=110
 exclude = jwst/extern,docs,jwst/associations,jwst/jwpsf,jwst/ramp_fitting, jwst/lib, jwst/fits_generator, jwst/exp_to_source, jwst/steps.py
 
 [tool:pytest]
-minversion = 3
+minversion = 3.6
 addopts = --ignore=build, -p no:warnings
 norecursedirs = .eggs build docs/_build relic jwst/timeconversion jwst/extern scripts
 asdf_schema_root = jwst/transforms/schemas jwst/datamodels/schemas
